Remove XML re-parsing from Player Abnormalities

+++ HunterPie/Core/Client/GStrings.cs
@@ -54,7 +54,7 @@ namespace HunterPie.Core {
             return Stage?.Attributes["Name"].Value;
         }
 
-        public static string GetAbnormalityByID(string Type, int ID, byte Stack) {
+        public static string GetAbnormalityByID(string Type, int ID, int Stack) {
             XmlNode Abnormality = Translations.SelectSingleNode($"//Strings/Abnormalities/Abnormality[@ID='{Type}_{ID:000}_{Stack:00}']");
             if (Abnormality == null) {
                 Abnormality = Translations.SelectSingleNode($"//Strings/Abnormalities/Abnormality[@ID='{Type}_{ID:000}_{0:00}']");
